ImageMap now accounts for screen density factor. Until now it always used density 1.0, which corresponds to Android MDPI, so areas and touch handling were done incorrectly on HDPI or higher resolutions.
Instead of using HashMap<Integer, Area> we now use SparseArray<Area> for a slight performance improvement.

When creating an ImageMap, new parameters can be defined in XML:
- ctc:fitImageToScreen
- ctc:mScaleFromOriginal
- ctc:maxSizeFactor (how big the pic scale factor can be; defaults to 1.5f)

The OnImageMapClickedHandler interface has been changed:
the onImageMapClicked() callback now accepts an area id as well as the ImageMap itself.
This has been done for the case of searching for something across several imagemaps.